% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differExp_continuous.R
\name{differExp_continuous}
\alias{differExp_continuous}
\title{Find differential expression genes or miRNAs from given expression data}
\usage{
differExp_continuous(se, class, log2 = FALSE, p_value.cutoff = 0.05)
}
\arguments{
\item{se}{\code{\link[SummarizedExperiment]{SummarizedExperiment}}
for input format.}

\item{class}{string. Choose one features from all rows of phenotype data.}

\item{log2}{logical, if this data hasn't been log2 transformed yet, this one
should be TRUE. Default is FALSE.}

\item{p_value.cutoff}{an numeric value indicating a threshold of p-value
for every genes or miRNAs (rows). Default is 0.05.}
}
\value{
data expression data in matrix format, with sample name in columns and
   gene symbol or miRNA name in rows.
}
\description{
This function will apply linear regression model to find differential expression
genes or miRNAs with continuous phenotype data,and then filter the genes or
miRNAs (rows) which have bigger p-value than cutoff.
}
\examples{
## Use the internal dataset
data("mirna", package = "anamiR", envir = environment())
data("pheno.mirna", package = "anamiR", envir = environment())

## SummarizedExperiment class
require(SummarizedExperiment)
mirna_se <- SummarizedExperiment(
 assays = SimpleList(counts=mirna),
 colData = pheno.mirna)

## Finding differential miRNA from miRNA expression data with lm
differExp_continuous(
    se = mirna_se, class = "Survival"
)

}
\seealso{
\code{\link[stats]{lm}} for fitting linear models.
}

