% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differExp_discrete.R
\name{differExp_discrete}
\alias{differExp_discrete}
\title{Find differential expression genes or miRNAs from given expression data}
\usage{
differExp_discrete(se, class, method = c("t.test", "limma", "wilcox.test",
  "DESeq"), limma.trend = FALSE, t_test.var = FALSE, log2 = FALSE,
  p_value.cutoff = 0.05, p_adjust.method = "BH", foldchange = 0.5)
}
\arguments{
\item{se}{\code{\link[SummarizedExperiment]{SummarizedExperiment}}
for input format.}

\item{class}{string. Choose one features from all rows of phenotype data.}

\item{method}{statistical method for finding differential genes or miRNAs,
including "t.test", "wilcox.test", "limma". Default is "t.test".}

\item{limma.trend}{logical, only matter when limma is chosen to be the method.
From function \code{\link[limma]{eBayes}}.}

\item{t_test.var}{logical, only matter when limma is chosen to be the method.
Whether to treat the two variances as being equal. From function
\code{\link[stats]{t.test}}}

\item{log2}{logical, if this data hasn't been log2 transformed yet, this one
should be TRUE. Default is FALSE.}

\item{p_value.cutoff}{an numeric value indicating a threshold of p-value
for every genes or miRNAs (rows). Default is 0.05.}

\item{p_adjust.method}{Correction method for multiple testing. (If you are
using DESeq for method, this param would not affect the result) From
function \code{\link[stats]{p.adjust}}. Default is "BH".}

\item{foldchange}{an numeric value indicating a threshold of foldchange (log2)
for every genes or miRNAs (rows). Default is 0.5.}
}
\value{
data expression data in matrix format, with sample name in columns and
   gene symbol or miRNA name in rows.
}
\description{
This function will apply one of three statistical methods, including t.test,
wilcox.test and limma, to find differential expression genes or miRNAs with,
discrete phenotype data, and then filter the genes or miRNAs (rows) which
have bigger p-value than cutoff.
}
\examples{
## Use the internal dataset
data("mirna", package = "anamiR", envir = environment())
data("pheno.mirna", package = "anamiR", envir = environment())

## SummarizedExperiment class
require(SummarizedExperiment)
mirna_se <- SummarizedExperiment(
 assays = SimpleList(counts=mirna),
 colData = pheno.mirna)

## Finding differential miRNA from miRNA expression data with t.test
mirna_d <- differExp_discrete(
   se = mirna_se,
   class = "ER",
   method = "t.test"
)

}
\seealso{
\code{\link[stats]{t.test}} for Student's t-Test;
   \code{\link[stats]{wilcox.test}} for Wilcoxon Rank Sum and Signed Rank
   Tests.
}

