% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egSymb.R
\docType{data}
\name{egSymb}
\alias{egSymb}
\title{A table with information of gene symbol and gene ID}
\format{A large matrix with 40784 rows and 2 columns:
\describe{
  \item{eg}{gene ID, in column}
  \item{sym}{gene symbol, in column}
  }}
\source{
\code{\link[gage]{gage}}
\url{https://bioconductor.org/packages/release/bioc/html/gage.html}
}
\usage{
egSymb
}
\value{
matrix
}
\description{
This table is originally from \code{\link[gage]{gage}},
including 40784 genes.
}
\keyword{datasets}
