% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEA_ana.R
\name{GSEA_ana}
\alias{GSEA_ana}
\title{This function will do GSEA analysis through the function
\code{\link[gage]{gage}}. After obtaining the ranking of
pathways, this function will choose the top five (default)
pathaways, and then find the related miRNAs based on their
gene set.}
\usage{
GSEA_ana(mrna, mirna, ref = NULL, samp = NULL, eg2sym = TRUE,
  pathway_num = 5)
}
\arguments{
\item{mrna}{differential expressed data in matrix format, with
sample name in columns and gene symbol in rows, which is generated by
\link{differExp_discrete} or \link{differExp_continuous}.}

\item{mirna}{differential expressed data in matrix format, with
sample name in columns and miRNA in rows. miRNA should
be miRBase 21 version now.}

\item{ref}{a numeric vector of column numbers for the reference
condition or phenotype (i.e. the control group) in the exprs
data matrix. Default ref = NULL, all columns are considered
as target experiments. from \code{\link[gage]{gage}}}

\item{samp}{a numeric vector of column numbers for the target
condition or phenotype (i.e. the experiment group) in the
exprs data matrix. Default samp = NULL, all columns other
than ref are considered as target experiments. from
\code{\link[gage]{gage}}}

\item{eg2sym}{logical. conversion between Entrez Gene IDs
and official gene symbols for human genes.}

\item{pathway_num}{The number of chosen pathways from the
result of GSEA analysis.}
}
\value{
list format containing both selected gene and miRNA
   expression data for each chosen pathway.
}
\description{
This function will do GSEA analysis through the function
\code{\link[gage]{gage}}. After obtaining the ranking of
pathways, this function will choose the top five (default)
pathaways, and then find the related miRNAs based on their
gene set.
}
\examples{

require(data.table)

## Load example data
aa <- system.file("extdata", "GSE19536_mrna.csv", package = "anamiR")
mrna <- fread(aa, fill = TRUE, header = TRUE)

bb <- system.file("extdata", "GSE19536_mirna.csv", package = "anamiR")
mirna <- fread(bb, fill = TRUE, header = TRUE)

## adjust data format
mirna_name <- as.matrix(mirna[, 1])
mrna_name <- as.matrix(mrna[, 1])
mirna <- mirna[, -1]
mrna <- mrna[, -1]
mirna <- data.matrix(mirna)
mrna <- data.matrix(mrna)
row.names(mirna) <- mirna_name
row.names(mrna) <- mrna_name

table <- GSEA_ana(mrna = mrna, mirna = mirna, pathway_num = 2)

}
\seealso{
\code{\link[gage]{gage}} for GSEA analysis.
}

