% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kegg.gs.R
\docType{data}
\name{kegg.gs}
\alias{kegg.gs}
\title{KEGG pathways with gene set information}
\format{A list with 177 KEGG pathways.}
\source{
\code{\link[gage]{gage}}
\url{https://bioconductor.org/packages/release/bioc/html/gage.html}
}
\usage{
kegg.gs
}
\value{
list
}
\description{
This table is originally from \code{\link[gage]{gage}},
including 177 pathways.
}
\keyword{datasets}

