% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_Differ.R
\name{multi_Differ}
\alias{multi_Differ}
\title{Find differential expression groups of each genes or miRNA from expression data}
\usage{
multi_Differ(se, class, anova_p_value = 0.05, post_hoc = c("scheffe.test",
  "duncan.test", "HSD.test"), post_hoc_p_value = 0.05)
}
\arguments{
\item{se}{\code{\link[SummarizedExperiment]{SummarizedExperiment}}
for input format.}

\item{class}{string. Choose one features from all rows of phenotype data.}

\item{anova_p_value}{an numeric value indicating a threshold of p-value from anova
for every genes or miRNAs (rows). Default is 0.05.}

\item{post_hoc}{post hoc test for anova, including  "scheffe.test", "HSD.test",
"duncan.test".}

\item{post_hoc_p_value}{an numeric value indicating a threshold of p-value from
post hoc test for every genes or miRNAs (rows). Default is 0.05.}
}
\value{
data.frame format with extra columns containing information about
   differential expressed groups among all.
}
\description{
This function will apply anova ,a statistical methods, for each gene or miRNA (row)
to find not only whether expression data of multiple groups differential expressed
or not, but also tell specifically two groups from all are differential expression.
}
\examples{
## Use the internal dataset
data("mirna", package = "anamiR", envir = environment())
data("pheno.mirna", package = "anamiR", envir = environment())

## SummarizedExperiment class
require(SummarizedExperiment)
mirna_se <- SummarizedExperiment(
 assays = SimpleList(counts=mirna),
 colData = pheno.mirna)

## Finding differential miRNA from miRNA expression data with anova
aov <- multi_Differ(se = mirna_se, class = "Subtype",
   post_hoc = "scheffe.test")

}
\seealso{
\code{\link[stats]{aov}} for fit an analysis of variance model.
}

