% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_pheno.mrna.R
\docType{data}
\name{pheno.mrna}
\alias{pheno.mrna}
\title{phenotype data of \code{\link{mrna}} about breast cancer}
\format{A data frame with 30 obs and 3 variables:
\describe{
  \item{ER}{estrogen receptor status}
  \item{Subtype}{breast cancer subtype}
  \item{Survival}{disease free survival time (months)}
 }}
\source{
NCBI GEO:
 \url{http://www.ncbi.nlm.nih.gov/geo/}
}
\usage{
pheno.mrna
}
\value{
data.frame
}
\description{
This phenotype dataset is originally from GSE19536. It contains
3 features in row and 30 samples in column.
}
\keyword{datasets}

