% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_table_pre.R
\docType{data}
\name{table_pre}
\alias{table_pre}
\title{A list with information of genes and miRNAs.}
\format{A large list with 2 gene expression files
 and 2 miRNAs files}
\source{
NCBI GEO:
 \url{http://www.ncbi.nlm.nih.gov/geo/}
}
\usage{
table_pre
}
\value{
list
}
\description{
This table is generated from \link{miR_converter},
including 4 files.
}
\keyword{datasets}
