% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEA_res.R
\name{GSEA_res}
\alias{GSEA_res}
\title{Pipeline of anamiR is applied to given output from \link{GSEA_ana}.}
\usage{
GSEA_res(table, pheno.data, class, DE_method = c("t.test", "limma",
  "wilcox.test", "DESeq"), limma.trend = FALSE, t_test.var = FALSE,
  log2 = FALSE, p_adjust.method = "BH", cor_cut = -0.3)
}
\arguments{
\item{table}{list format containing both selected gene and miRNA
expression data for each chosen pathway. output of \link{GSEA_ana}}

\item{pheno.data}{phenotype data.}

\item{class}{string. Choose one features from all rows of phenotype data.}

\item{DE_method}{statistical method for finding differential genes or miRNAs,
including "t.test", "wilcox.test", "limma". Default is "t.test".}

\item{limma.trend}{logical, only matter when limma is chosen to be the method.
From function \code{\link[limma]{eBayes}}.}

\item{t_test.var}{logical, only matter when limma is chosen to be the method.
Whether to treat the two variances as being equal. From function
\code{\link[stats]{t.test}}}

\item{log2}{logical, if this data hasn't been log2 transformed yet, this one
should be TRUE. Default is FALSE.}

\item{p_adjust.method}{Correction method for multiple testing. (If you are
using DESeq for method, this param would not affect the result) From
function \code{\link[stats]{p.adjust}}. Default is "BH".}

\item{cor_cut}{an numeric value indicating a threshold of correlation
coefficient for every potential miRNA-genes interactions. Default is -0.3,
however, if no interaction pass the threshold, this function would add
0.2 value in threshold until at least one interaction passed the threshold.}
}
\value{
list format containing matrix for each chosen pathway.
   The format of matrix is like the output from \link{negative_cor}.
}
\description{
This function will use \link{differExp_discrete} and
\link{negative_cor} to do the deeper analysis of given data which
is from \link{GSEA_ana}.
}
\examples{
## Load example data

require(data.table)

cc <- system.file("extdata", "pheno_data.csv", package = "anamiR")
pheno.data <- fread(cc, fill = TRUE, header = TRUE)

## adjust data format
pheno_name <- pheno.data[["Sample"]]
pheno.data <- pheno.data[, -1]
pheno.data <- as.matrix(pheno.data)
row.names(pheno.data) <- pheno_name
data(table_pre)

result <- GSEA_res(table = table_pre, pheno.data = pheno.data,
 class = "ER", DE_method = "limma", cor_cut = 0)

}
\seealso{
\link{differExp_discrete} and \link{negative_cor}.
}
